Moved add_feature_flag out of integrations
